only run database-tests once
